# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/inventory/locations": {
      "get": {
        "operationId": "InventoryController_locations_v1",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "itmarket_staff_access": [] }],
        "tags": ["inventory"]
      },
      "post": {
        "operationId": "InventoryController_createLocation_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LocationDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "itmarket_staff_access": [] }],
        "tags": ["inventory"]
      }
    },
    "/api/v1/inventory/locations/{id}": {
      "patch": {
        "operationId": "InventoryController_updateLocation_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LocationDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "security": [{ "itmarket_staff_access": [] }],
        "tags": ["inventory"]
      }
    },
    "/api/v1/inventory/balances": {
      "get": {
        "operationId": "InventoryController_balances_v1",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "itmarket_staff_access": [] }],
        "tags": ["inventory"]
      }
    },
    "/api/v1/inventory/movements": {
      "get": {
        "operationId": "InventoryController_movements_v1",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "itmarket_staff_access": [] }],
        "tags": ["inventory"]
      }
    },
    "/api/v1/inventory/receipts": {
      "post": {
        "operationId": "InventoryController_receipt_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MovementDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "itmarket_staff_access": [] }],
        "tags": ["inventory"]
      }
    },
    "/api/v1/inventory/adjustments": {
      "post": {
        "operationId": "InventoryController_adjustment_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AdjustmentDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "itmarket_staff_access": [] }],
        "tags": ["inventory"]
      }
    },
    "/api/v1/inventory/transfers": {
      "post": {
        "operationId": "InventoryController_transfer_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransferDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "itmarket_staff_access": [] }],
        "tags": ["inventory"]
      }
    },
    "/api/v1/inventory/reconciliation": {
      "get": {
        "operationId": "InventoryController_reconcile_v1",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "itmarket_staff_access": [] }],
        "tags": ["inventory"]
      }
    },
...
    "/api/v1/reports/inventory/movements": {
      "get": {
        "operationId": "ReportsController_movements_v1",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "itmarket_staff_access": [] }],
        "tags": ["reports"]
      }
    },
...
  },
  "info": {
    "title": "ITMarket API",
    "description": "ITMarket service API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [{ "url": "/api/v1" }],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "itmarket_customer_session",
        "description": "Opaque, HTTP-only customer session cookie"
      }
    },
    "schemas": {
...
      "LocationDto": { "type": "object", "properties": {} },
      "MovementDto": { "type": "object", "properties": {} },
      "AdjustmentDto": { "type": "object", "properties": {} },
      "TransferDto": { "type": "object", "properties": {} },
...
    }
  }
}